refractored functions
